Adds fav/boost class view
