+++ bookwyrm/tests/test_outgoing.py
@@ -430,112 +430,3 @@ class Outgoing(TestCase):
         self.assertFalse(models.Review.objects.filter(
             book=self.book, user=self.local_user
         ).exists())
-
-
-    def test_handle_delete_status(self):
-        ''' marks a status as deleted '''
-        status = models.Status.objects.create(
-            user=self.local_user, content='hi')
-        self.assertFalse(status.deleted)
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            outgoing.handle_delete_status(self.local_user, status)
-        status.refresh_from_db()
-        self.assertTrue(status.deleted)
-
-
-    def test_handle_favorite(self):
-        ''' create and broadcast faving a status '''
-        status = models.Status.objects.create(
-            user=self.local_user, content='hi')
-
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            outgoing.handle_favorite(self.remote_user, status)
-        fav = models.Favorite.objects.get()
-        self.assertEqual(fav.status, status)
-        self.assertEqual(fav.user, self.remote_user)
-
-        notification = models.Notification.objects.get()
-        self.assertEqual(notification.notification_type, 'FAVORITE')
-        self.assertEqual(notification.user, self.local_user)
-        self.assertEqual(notification.related_user, self.remote_user)
-
-
-    def test_handle_unfavorite(self):
-        ''' unfav a status '''
-        status = models.Status.objects.create(
-            user=self.local_user, content='hi')
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            outgoing.handle_favorite(self.remote_user, status)
-
-        self.assertEqual(models.Favorite.objects.count(), 1)
-        self.assertEqual(models.Notification.objects.count(), 1)
-
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            outgoing.handle_unfavorite(self.remote_user, status)
-        self.assertEqual(models.Favorite.objects.count(), 0)
-        self.assertEqual(models.Notification.objects.count(), 0)
-
-
-    def test_handle_boost(self):
-        ''' boost a status '''
-        status = models.Status.objects.create(
-            user=self.local_user, content='hi')
-
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            outgoing.handle_boost(self.remote_user, status)
-
-        boost = models.Boost.objects.get()
-        self.assertEqual(boost.boosted_status, status)
-        self.assertEqual(boost.user, self.remote_user)
-        self.assertEqual(boost.privacy, 'public')
-
-        notification = models.Notification.objects.get()
-        self.assertEqual(notification.notification_type, 'BOOST')
-        self.assertEqual(notification.user, self.local_user)
-        self.assertEqual(notification.related_user, self.remote_user)
-        self.assertEqual(notification.related_status, status)
-
-    def test_handle_boost_unlisted(self):
-        ''' boost a status '''
-        status = models.Status.objects.create(
-            user=self.local_user, content='hi', privacy='unlisted')
-
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            outgoing.handle_boost(self.remote_user, status)
-
-        boost = models.Boost.objects.get()
-        self.assertEqual(boost.privacy, 'unlisted')
-
-    def test_handle_boost_private(self):
-        ''' boost a status '''
-        status = models.Status.objects.create(
-            user=self.local_user, content='hi', privacy='followers')
-
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            outgoing.handle_boost(self.remote_user, status)
-        self.assertFalse(models.Boost.objects.exists())
-
-    def test_handle_boost_twice(self):
-        ''' boost a status '''
-        status = models.Status.objects.create(
-            user=self.local_user, content='hi')
-
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            outgoing.handle_boost(self.remote_user, status)
-            outgoing.handle_boost(self.remote_user, status)
-        self.assertEqual(models.Boost.objects.count(), 1)
-
-
-    def test_handle_unboost(self):
-        ''' undo a boost '''
-        status = models.Status.objects.create(
-            user=self.local_user, content='hi')
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            outgoing.handle_boost(self.remote_user, status)
-
-        self.assertEqual(models.Boost.objects.count(), 1)
-        self.assertEqual(models.Notification.objects.count(), 1)
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            outgoing.handle_unboost(self.remote_user, status)
-        self.assertEqual(models.Boost.objects.count(), 0)
-        self.assertEqual(models.Notification.objects.count(), 0)
